(refactor: directory) 투어 디렉터리 제거
