+++ src/main/java/com/weatherwhere/airservice/service/airrealtime/GetTmXYAndStationService.java
@@ -1,4 +1,4 @@
-package com.weatherwhere.airservice.service;
+package com.weatherwhere.airservice.service.airrealtime;
 
 public interface GetTmXYAndStationService {
     // 경도 x, 위도 y로 측정소명 받기
