+++ src/main/java/com/weatherwhere/airservice/service/airrealtime/TourRankAirDataService.java
@@ -1,7 +1,7 @@
-package com.weatherwhere.airservice.service.tour;
+package com.weatherwhere.airservice.service.airrealtime;
 
 import com.weatherwhere.airservice.dto.ResultDTO;
-import com.weatherwhere.airservice.dto.tour.TourAirRealTimeDataDTO;
+import com.weatherwhere.airservice.dto.airrealtime.TourAirRealTimeDataDTO;
 
 public interface TourRankAirDataService {
     ResultDTO<TourAirRealTimeDataDTO> getRealTimeAirDBData(Double x, Double y);
